[FEATURE] Updated views to be device responsive
