+++ src/layouts/dashboard.jsx
@@ -1,14 +1,24 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import {
-  Hidden, withWidth,
+  Hidden, withWidth, List, ListItem,
 } from '@material-ui/core';
-import MobileView from '../templates/mobileDashBoard';
-import DesktopView from '../templates/desktopDashboard';
+import MobileView from '../templates/DashboardSM';
+import DesktopView from '../templates/DashBoardLG';
+import Card from '../templates/Card';
 
 const useStyles = makeStyles(() => ({
   root: {
     flexGrow: 1,
+  },
+  SMCard: {
+    minWidth: 280,
+    width: '90vw',
+    opacity: 0.9,
+  },
+  LGCard: {
+    width: '75vw',
+    opacity: 0.9,
   },
   container: {
     display: 'flex',
@@ -16,21 +26,37 @@ const useStyles = makeStyles(() => ({
   },
 }));
 
-function BreakpointOnly() {
+function Dashboard() {
   const classes = useStyles();
 
   return (
     <div className={classes.root}>
       <div className={classes.container}>
         <Hidden only={['lg', 'xl']}>
-          <DesktopView />
+          <MobileView content={[...new Array(7)]
+            .map((index) => (
+              <List>
+                <ListItem key={index}>
+                  <Card classAttr={classes.SMCard} button />
+                </ListItem>
+              </List>
+            ))}
+          />
         </Hidden>
         <Hidden only={['sm', 'xs', 'md']}>
-          <MobileView />
+          <DesktopView content={[...new Array(7)]
+            .map((index) => (
+              <List>
+                <ListItem key={index}>
+                  <Card classAttr={classes.LGCard} button />
+                </ListItem>
+              </List>
+            ))}
+          />
         </Hidden>
       </div>
     </div>
   );
 }
 
-export default withWidth()(BreakpointOnly);
+export default withWidth()(Dashboard);
