--- src/templates/mobile/Drawer.jsx
@@ -1,87 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { makeStyles, useTheme } from '@material-ui/core/styles';
-import {
-  Drawer, IconButton, Divider, List,
-  ListItem, ListItemText, ListItemIcon,
-} from '@material-ui/core';
-import {
-  ChevronLeft, ChevronRight, AccountCircle, Home,
-} from '@material-ui/icons';
-
-
-const drawerWidth = 150;
-
-const useStyles = makeStyles((theme) => ({
-  drawer: {
-    width: drawerWidth,
-    flexShrink: 0,
-  },
-  drawerPaper: {
-    width: drawerWidth,
-  },
-  drawerHeader: {
-    display: 'flex',
-    alignItems: 'center',
-    padding: theme.spacing(0, 1),
-    // necessary for content to be below app bar
-    ...theme.mixins.toolbar,
-    justifyContent: 'flex-end',
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(3),
-    transition: theme.transitions.create('margin', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    marginLeft: -drawerWidth,
-  },
-  contentShift: {
-    transition: theme.transitions.create('margin', {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-    marginLeft: 0,
-  },
-}));
-
-function Sidebar({ closeState, drawerCloser }) {
-  const classes = useStyles();
-  const theme = useTheme();
-  return (
-    <Drawer
-      className={classes.drawer}
-      variant="persistent"
-      anchor="left"
-      open={closeState}
-      classes={{
-        paper: classes.drawerPaper,
-      }}
-    >
-      <div className={classes.drawerHeader}>
-        <IconButton onClick={drawerCloser}>
-          {theme.direction === 'ltr' ? <ChevronLeft /> : <ChevronRight />}
-        </IconButton>
-      </div>
-      <Divider />
-      <List>
-        {[{ text: 'Home', icon: <Home /> }, { text: 'Profile', icon: <AccountCircle /> }]
-          .map(({ text, icon }) => (
-            <ListItem button key={text}>
-              <ListItemIcon>{icon}</ListItemIcon>
-              <ListItemText primary={text} />
-            </ListItem>
-          ))}
-      </List>
-      <Divider />
-    </Drawer>
-  );
-}
-
-Sidebar.propTypes = {
-  closeState: PropTypes.bool.isRequired,
-  drawerCloser: PropTypes.func.isRequired,
-};
-
-export default Sidebar;
